wipe-module: move tests
